Repair code quality and Add UG  Add a user guide and fix the code quality errors
