+++ src/main/java/aquadem/Event.java
@@ -1,4 +1,4 @@
-package Aquadem;
+package aquadem;
 
 import java.io.Serializable;
