+++ src/main/java/aquadem/Task.java
@@ -1,9 +1,7 @@
-package Aquadem;
+package aquadem;
 
 import java.io.Serializable;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.time.Month;
 
 /**
  * Class to encapsulate the behaviour of task in the chatbots tasklist.
@@ -49,7 +47,6 @@ public class Task implements Serializable {
      * @param date
      */
     public void setDate(LocalDateTime date) {
-        assert date.isAfter(LocalDateTime.now()) : "Too old";
         this.taskDate = date;
     }
