+++ src/main/java/javaFx/DialogBox.java
@@ -58,4 +58,5 @@ public class DialogBox extends HBox {
         db.flip();
         return db;
     }
+
 }
