+++ src/test/java/aquadem/ExecutionTest.java
@@ -1,4 +1,4 @@
-package Aquadem;
+package aquadem;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDateTime;
